+++ sdk/appconfiguration/azure-appconfiguration/azure/appconfiguration/_azure_appconfiguration_client.py
@@ -37,7 +37,9 @@ from ._utils import (
     get_label_filter,
     parse_connection_string,
 )
+
 from ._sync_token import SyncTokenPolicy
+from ._audience_policy import AudiencePolicy
 
 
 class AzureAppConfigurationClient:
@@ -66,6 +68,10 @@ class AzureAppConfigurationClient:
         credential_scopes = [f"{base_url.strip('/')}/.default"]
         self._sync_token_policy = SyncTokenPolicy()
 
+        audience_policy = AudiencePolicy(kwargs.get("credential_scopes", None))
+
+        per_call_policies = [self._sync_token_policy, audience_policy]
+
         if isinstance(credential, AzureKeyCredential):
             id_credential = kwargs.pop("id_credential")
             kwargs.update(
@@ -85,7 +91,7 @@ class AzureAppConfigurationClient:
             )
         # mypy doesn't compare the credential type hint with the API surface in patch.py
         self._impl = AzureAppConfigurationClientGenerated(
-            base_url, credential, per_call_policies=self._sync_token_policy, **kwargs  # type: ignore[arg-type]
+            base_url, credential, per_call_policies=per_call_policies, **kwargs  # type: ignore[arg-type]
         )
 
     @classmethod